Seperated the pronoun_picker from the driver
